fix: Make generating docs non destructive (#323)
#### Summary

Before this change, after the first time a description was written to the string it would be deleted. Because all of the `schema` objects are passed by reference, this means that it would be permanently deleted if the initial object is passed in as a struct and not marshaled/unmarshaled.

This PR removes that specific destructive aspect and instead overwrites the `Comments` attribute which we do not use. This change doesn't impact any existing code as all instances still use the `Generate` rather than the `GenerateFromSchema`